xaga: Patch Goodix fingerprint HAL to ignore conflicting events
This will fix the long standing issue of the screenshot button
combination not working properly in AOSP for Goodix sensors.

The patch will replace the first instruction of the
'gf_hal_send_key_event' function with a ret so that the touch
events which conflict with AOSP do not get reported to userspace.

Change-Id: I9e15a46b519fd7f75e39af4690c271aa63af4790
Signed-off-by: ChrisCatto <dontforgetme047@gmail.com>
Signed-off-by: priiii1808 <priyanshusinghal0818@gmail.com>